Separated DroppedEvents and extracted common styles between DroppedClass and DroppedEvents  to a separate file
